backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
psycopg[binary]==3.1.16
psycopg2-binary==2.9.10
pydantic-settings>=2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.7
python-dotenv==1.0.0
pytest==7.4.4
httpx>=0.27.0,<1.0.0
asgi-lifespan==2.1.0
tenacity==8.2.3
uuid==1.30
sqlalchemy==2.0.40
alembic==1.15.2
email-validator==2.2.0
slowapi==0.1.9
openai==1.78.0
tiktoken==0.9.0
openai-agents==0.0.14

